Establish Tutorial Layout
Each and every tutorial and topic I would like to discuss has been laid out. This gives a neat overview of what I have left to discuss and serves as a kind of todo list, as well as giving my mentor on this project an idea of what I want- and have left to cover.